+++ lib/flutter_carousel_widget.dart
@@ -8,9 +8,9 @@ export 'package:flutter_carousel_widget/src/helpers/flutter_carousel_state.dart'
 export 'package:flutter_carousel_widget/src/helpers/flutter_expandable_carousel_controller.dart';
 export 'package:flutter_carousel_widget/src/helpers/flutter_expandable_carousel_options.dart';
 export 'package:flutter_carousel_widget/src/helpers/flutter_expandable_carousel_state.dart';
+export 'package:flutter_carousel_widget/src/helpers/slide_indicator_options_model.dart';
 export 'package:flutter_carousel_widget/src/indicators/circular_slide_indicator.dart';
 export 'package:flutter_carousel_widget/src/indicators/circular_static_indicator.dart';
 export 'package:flutter_carousel_widget/src/indicators/circular_wave_slide_indicator.dart';
-export 'package:flutter_carousel_widget/src/indicators/models/slide_indicator_options_model.dart';
 export 'package:flutter_carousel_widget/src/indicators/sequential_fill_indicator.dart';
 export 'package:flutter_carousel_widget/src/indicators/slide_indicator.dart';